DFE-1163 fully moving over to versioning model
